Update webpack sass loader to properly find node_modules, for material design
I was trying to import an MD component and ran into an error where the sass loader couldnt find the npm folder. I tried working around this by adding a ~ in front, but then ran into the issue of being unable to find any of the other dependencies required by the first file. This was resolved by editing the webpack config file to look at node modules (it looked like it was already doing this...but maybe not?).